requirements.txt
aiofiles==23.2.1
aiohttp==3.9.0
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
sentry-sdk==1.37.0
aiosqlite==0.3.0
pymongo==4.6.0
multidict==6.0.4
